organized components into auth, pages, and ui folders
