+++ frontend/src/app.css
@@ -8,12 +8,6 @@
     margin-top: 50px;
 }
 
-#copyright {
-	height: 75px;
-	text-align: center;
-	margin-top: auto;
-}
-
 .bg-myBlue {
     background-color: RoyalBlue;
 }
