+++ frontend/src/components/auth/register/password-validator.ts
