+++ frontend/src/components/auth/register/register.css
