+++ frontend/src/components/auth/register/register.tsx
