+++ frontend/src/components/pages/about.tsx
