+++ frontend/src/components/pages/view_notes/view_notes.css
