+++ frontend/src/components/ui/footer/footer.css
@@ -1,0 +1,5 @@
+#copyright {
+	height: 75px;
+	text-align: center;
+	margin-top: auto;
+}
