+++ frontend/src/components/ui/footer/footer.tsx
@@ -1,0 +1,11 @@
+/* Navigation Bar */
+
+import "./footer.css";
+
+export default function Footer() {
+	return (
+		<footer id="copyright">
+			<p> &copy; {(new Date().getFullYear())} Salmaan Saeed </p>
+		</footer>
+	);
+}
