+++ frontend/src/components/ui/navigation_bar/navigation_bar.css
